# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Notly projesinin backend API dokümantasyonu.",
        "title": "Notly API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/schedule/blocked-slots": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "GetBlockedSlots",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date (YYYY-MM-DD)",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_M1ralai_notly-api_internal_modules_schedule_dto.BlockedSlotResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "CreateBlockedSlot",
                "parameters": [
                    {
                        "description": "Create Blocked Slot",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_notly-api_internal_modules_schedule_dto.CreateBlockedSlotRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_notly-api_internal_modules_schedule_dto.BlockedSlotResponse"
                        }
                    }
                }
            }
        },
        "/api/schedule/blocked-slots/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "DeleteBlockedSlot",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/schedule/check-conflict": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "CheckConflict",
                "parameters": [
                    {
                        "description": "Conflict Check Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_notly-api_internal_modules_schedule_dto.ConflictCheckRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_notly-api_internal_modules_schedule_dto.ConflictResponse"
                        }
                    }
                }
            }
        },
        "/api/schedule/free-slots": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "GetFreeSlots",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date (YYYY-MM-DD)",
                        "name": "date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Duration in minutes (minimum 15)",
                        "name": "duration",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_M1ralai_notly-api_internal_modules_schedule_dto.TimeSlotResponse"
                            }
                        }
                    }
                }
            }
        },
        "/api/schedule/generate-events": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "GenerateEvents",
                "parameters": [
                    {
                        "description": "Generate Events Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_notly-api_internal_modules_schedule_dto.GenerateEventsRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_M1ralai_notly-api_internal_modules_schedule_dto.GenerateEventsResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_M1ralai_notly-api_internal_modules_schedule_dto.BlockedSlotResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "end_datetime": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_flexible": {
                    "type": "boolean"
                },
                "reason": {
                    "type": "string"
                },
                "source_type": {
                    "type": "string"
                },
                "start_datetime": {
                    "type": "string"
                }
            }
        },
        "github_com_M1ralai_notly-api_internal_modules_schedule_dto.ConflictCheckRequest": {
            "type": "object",
            "required": [
                "end",
                "start"
            ],
            "properties": {
                "end": {
                    "type": "string"
                },
                "start": {
                    "type": "string"
                }
            }
        },
        "github_com_M1ralai_notly-api_internal_modules_schedule_dto.ConflictResponse": {
            "type": "object",
            "properties": {
                "has_conflict": {
                    "type": "boolean"
                },
                "reason": {
                    "type": "string"
                },
                "suggestions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_M1ralai_notly-api_internal_modules_schedule_dto.TimeSlotResponse"
                    }
                }
            }
        },
        "github_com_M1ralai_notly-api_internal_modules_schedule_dto.CreateBlockedSlotRequest": {
            "type": "object",
            "required": [
                "end_datetime",
                "reason",
                "start_datetime"
            ],
            "properties": {
                "end_datetime": {
                    "type": "string"
                },
                "is_flexible": {
                    "type": "boolean"
                },
                "reason": {
                    "type": "string"
                },
                "source_id": {
                    "type": "integer"
                },
                "source_type": {
                    "type": "string"
                },
                "start_datetime": {
                    "type": "string"
                }
            }
        },
        "github_com_M1ralai_notly-api_internal_modules_schedule_dto.GenerateEventsRequest": {
            "type": "object",
            "required": [
                "day_of_week",
                "end_time",
                "semester_end_date",
                "semester_start_date",
                "start_time",
                "title"
            ],
            "properties": {
                "course_schedule_id": {
                    "type": "integer"
                },
                "day_of_week": {
                    "type": "string",
                    "enum": [
                        "Monday",
                        "Tuesday",
                        "Wednesday",
                        "Thursday",
                        "Friday",
                        "Saturday",
                        "Sunday"
                    ]
                },
                "end_time": {
                    "type": "string"
                },
                "exclude_dates": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "location": {
                    "type": "string"
                },
                "semester_end_date": {
                    "type": "string"
                },
                "semester_start_date": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "github_com_M1ralai_notly-api_internal_modules_schedule_dto.GenerateEventsResponse": {
            "type": "object",
            "properties": {
                "blocked_slots_created": {
                    "type": "integer"
                },
                "events_generated": {
                    "type": "integer"
                }
            }
        },
        "github_com_M1ralai_notly-api_internal_modules_schedule_dto.TimeSlotResponse": {
            "type": "object",
            "properties": {
                "duration_minutes": {
                    "type": "integer"
                },
                "end": {
                    "type": "string"
                },
                "start": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}